requirements.txt
# Streamlit for web interface
streamlit>=1.28.0

# OpenAI and Azure OpenAI
openai>=1.3.0

# Vector database
chromadb>=0.4.15

# LangChain components
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-chroma>=0.1.0

# Document processing
GitPython>=3.1.40
python-dotenv>=1.0.0

# Text processing and utilities
tiktoken>=0.5.0
numpy>=1.24.0
pandas>=2.0.0

# HTTP requests
requests>=2.31.0

# File handling and text processing
chardet>=5.0.0
python-magic>=0.4.27

# Additional document loaders
unstructured>=0.10.0
python-docx>=0.8.11
PyPDF2>=3.0.0
markdown>=3.5.0

# Optional: Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0